server/requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
backoff==1.10.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.1.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
pylint==2.6.0
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
spotipy==2.17.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
